+++ config/bootstrap.php
@@ -20,7 +20,7 @@ if (!defined('DATABASE_BACKUP')) {
 
 //Sets the redirect to `/dev/null`. This string can be concatenated to shell commands
 if (!defined('REDIRECT_TO_DEV_NULL')) {
-    define('REDIRECT_TO_DEV_NULL', isWin() ? ' 2>nul' : ' 2>/dev/null');
+    define('REDIRECT_TO_DEV_NULL', is_win() ? ' 2>nul' : ' 2>/dev/null');
 }
 
 //Binaries